Do not read exchange & queue before deleting, simply delete
Doing a wread operation requires a table write lock. Since a delete
requires a table write lock as well, we have 2 write locks instead of 1.
When Mnesia is under load, this can result in many lock collisions and
slow down the entire set of operations. We think that it's better to
leverage fast SSDs / NVMEs and defer the Mnesia disk log write
optimisation to the filesystem.

For more context, see #1513

Partner-in-crime @essen