Wait at most 5 secs for a node to reply to rabbit_node_monitor:partitions/0
When rabbit_node_monitor that runs on an alive node is busy cleaning
after another node goes down, the process might not respond for a long
time.

Since the Management API calls rabbit_node_monitor:partitions/0 and
blocks until this function returns, the Management UI will fail to load
for a really long time - clients will timeout before this function
returns.

This change will make rabbit_node_monitor:partitions/0 timeout after 5
seconds so that the Management API returns in a timely manner and the
Management UI blocks for an extra 5 seconds at most. The implications of
this will result in node metrics being outdated, which is not ideal, but
at least the users will have some feedback and will be able to perform
action via the Management UI / API.

For more context, see #1513

Partner-in-crime: @essen